requirements.txt
paho-mqtt==1.6.1
Flask==2.2.3
Flask-SocketIO==5.3.2
Flask-MQTT==1.1.1
Werkzeug==2.2.3
python-socketio==5.7.2
